abstracted main page functions to separate file
